Remove old changes to locator sources. Change all widget headers to h1 for consistency.
